Fix: 기존에 브라우저 창 끄기 핸들러 커스텀훅 생성 후 모든 페이지에 구현완료
